:construction: Move leaderboard logic into its own component
For better separation of concerns, composability and overall maintenance reasons. Motivated by the need to update it to use apollo hooks and garden.
