+++ app/javascript/pages/Dashboard/index.js
@@ -1,20 +1,16 @@
-import React, { Component } from 'react'
+import React from 'react'
 import { graphql } from 'react-apollo'
 import { NetworkStatus } from 'apollo-client'
 import { connect } from 'react-redux'
 import * as R from 'ramda'
 import moment from 'moment'
 
-import OfficeFilter from 'components/OfficeFilter'
 import Layout from 'components/Layout'
 import Loading from 'components/LoadingIcon'
-import NamedAvatar from 'components/NamedAvatar'
 import ProgressCircle from 'components/ProgressCircle'
-
-import { changeDashboardOfficeFilter } from 'actions'
+import Leaderboard from './Leaderboard'
 
 import MilestoneQuery from './milestoneQuery.gql'
-import LeaderboardQuery from './leaderboardQuery.gql'
 import s from './main.css'
 
 const milestones = [
@@ -29,9 +25,6 @@ const milestoneBarChunk = milestoneBarWidth / milestones.length
 // The server expects seconds since epoch, not milliseconds
 const startOfYear = Math.floor(moment().startOf('year') / 1000)
 const nowInSec = moment().unix()
-
-const leaderBoardSize = 10
-const leaderBoardSort = 'HOURS_DESC'
 
 const selectMilestone = hours => R.find(m => hours < m.hours)(milestones) || R.last(milestones)
 
@@ -117,64 +110,13 @@ const milestoneLabelStyling = (item, milestone, user) => {
   return R.join(' ', classes)
 }
 
-const LeaderboardContainer = ({
-  data: { networkStatus, volunteers, offices, currentUser },
-  dashboardOfficeFilter,
-  changeDashboardOfficeFilter,
-}) =>
-  networkStatus === NetworkStatus.loading ? (
-    <Loading />
-  ) : (
-    <div className={s.leaderboard}>
-      <div className={s.leaderboardHeader}>
-        <div className={s.topVolunteers}>Top Volunteers</div>
-        <OfficeFilter
-          offices={offices}
-          value={dashboardOfficeFilter.value === 'current' ? currentUser.office.id : dashboardOfficeFilter.value}
-          onChange={changeDashboardOfficeFilter}
-        />
-      </div>
-      {volunteers.map((user, i) => (
-        <div className={s.leaderboardUser} key={`user-${i}`}>
-          <NamedAvatar image={user.photo} name={user.name} subtitle={user.group} />
-          <span className={s.leaderboardHours}>{user.hours} hours</span>
-        </div>
-      ))}
-    </div>
-  )
-
 const mapStateToProps = (state, _ownProps) => {
-  const { dashboardOfficeFilter } = state.model
   const { locationBeforeTransitions } = state.routing
 
   return {
-    dashboardOfficeFilter,
     locationBeforeTransitions,
   }
 }
-
-const leaderboardWithData = graphql(LeaderboardQuery, {
-  options: ({ dashboardOfficeFilter }) => {
-    const variables = {
-      count: leaderBoardSize,
-      sortBy: leaderBoardSort,
-      after: startOfYear,
-      before: nowInSec,
-    }
-
-    variables.officeId = dashboardOfficeFilter.value
-
-    return {
-      variables,
-      fetchPolicy: 'cache-and-network',
-    }
-  },
-})
-const leaderboardWithActions = connect(
-  mapStateToProps,
-  { changeDashboardOfficeFilter }
-)
-const Leaderboard = leaderboardWithActions(leaderboardWithData(LeaderboardContainer))
 
 const Dashboard = ({ data: { networkStatus, currentUser }, locationBeforeTransitions }) => {
   if (networkStatus === NetworkStatus.loading) {
